+++ src/OutputLayer.cpp
@@ -4,10 +4,9 @@
 
 using namespace std;
 
-OutputLayer::OutputLayer(int size)
-    : Layer(size)
-{
-    for (int i = 0; i < size; i++) {
+OutputLayer::OutputLayer(unsigned long size)
+        : Layer(size) {
+    for (unsigned long i = 0; i < size; i++) {
         this->m_neurons.push_back(new OutputNeuron());
     }
 }
